Added delivery page

+++ src/MedicationManager.BusinessLogic/Stock/Dtos/BaseStockDto.cs
@@ -8,8 +8,8 @@ namespace MedicationManager.BusinessLogic.Stock.Dtos
         public DateTime Date { get; set; }
         public string MedicationId { get; set; }
         public int Count { get; set; }
-        public int PricePerItem { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal PricePerItem { get; set; }
+        public decimal TotalPrice { get; set; }
 
         public MedicationDto Medication { get; set; }
     }
